Add pseudo-instruction for large immediate materialization and an expansion pass.
The pseudo-instruction combines the previous 'mvui'/'ori' pair and
defines all bits of a CPUReg, so we can set the isRematerializable
bit on it.  This means that it can be hoisted out of loops by the
MachineLICM pass, which is nearly always a good thing.

We can use the ExpandPseudoInsts pass for other things later; there
are other chains of instructions that materialize some complex value
into a single register, but are best treated as a single unit until
late in the compilation process.

Fixes bug 93.